TutoFlask/requirements.txt
blinker==1.9.0
click==8.3.0
Flask==3.1.2
flask-bootstrap5==0.1.dev1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
python-dotenv==1.1.1
PyYAML==6.0.3
setuptools==80.9.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
Werkzeug==3.1.3
wheel==0.45.1
WTForms==3.2.1
